split "database.py" into separate files
